P3NewUI/weavestream-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
pydantic
